made leave work for all players

+++ src/main/java/ch/uzh/ifi/seal/soprafs20/service/LobbyService.java
@@ -216,17 +216,43 @@ public class LobbyService
         //todo fix
         Player player = playerService.getPlayerById(playerId);
         Lobby lobby = this.getLobbyById(lobbyId);
+        Player creator = lobby.getCreator();
         Game game = lobby.getGame();
         List<Clue> clues = player.getClues();
         for(Clue clue:clues){
         	game.deleteClue(clue);
             clue.setPlayer(null);
         }
-        lobby.leave(player);
-        lobby = lobbyRepository.save(lobby);
-        playerService.deletePlayer(player);
-        this.setNewPlayersStatus(lobby.getPlayers(), PlayerStatus.END_OF_TURN, PlayerStatus.END_OF_TURN);
-        lobbyRepository.flush();        
+        boolean guesserSet = false;
+        if((creator.getId() == playerId)||(player.getRole() == PlayerRole.GUESSER))
+        {
+        	Set<Player> players = lobby.getPlayers();
+        	for(Player candidatPlayer : players)
+        	{
+        		if((candidatPlayer.getId() != lobby.getCreator().getId())&&(creator.getId() == playerId))
+        		{
+        			lobby.setCreator(candidatPlayer);
+
+        		}
+        		if((candidatPlayer.getId() != playerId)&&(guesserSet==false))
+        		{
+        			candidatPlayer.setRole(PlayerRole.GUESSER);
+        			guesserSet = true;
+        		}
+        	}
+        }
+        if(lobby.getCreator().getId() == playerId)
+        {
+        	endLobby(lobbyId,userService.getUserByID(playerId));
+        }
+        else
+        {
+	       lobby.leave(player);
+	       lobby = lobbyRepository.save(lobby);
+	       playerService.deletePlayer(player);
+	       this.setNewPlayersStatus(lobby.getPlayers(), PlayerStatus.END_OF_TURN, PlayerStatus.END_OF_TURN);
+	       lobbyRepository.flush();  
+        }
     }
 
     /**
